+++ src/graph/planner/match/MatchPathPlanner.h
@@ -74,16 +74,6 @@ class MatchPathPlanner final {
                         SubPlan& subplan,
                         std::unordered_set<std::string>& nodeAliasesSeenInPattern);
 
-  // Project all named alias.
-  // TODO: Might not neccessary
-  Status projectColumnsBySymbols(QueryContext* qctx, Path& path, SubPlan& plan);
-
-  YieldColumn* buildVertexColumn(ObjectPool* pool, const std::string& alias) const;
-
-  YieldColumn* buildEdgeColumn(ObjectPool* pool, EdgeInfo& edge) const;
-
-  YieldColumn* buildPathColumn(Expression* pathBuild, const std::string& alias) const;
-
  private:
   Expression* initialExpr_{nullptr};
 };
